@@ -1,22 +1,53 @@
 const glob = require('fast-glob');
 const fs = require('fs');
 const assert = require('assert');
+const chalk = require('chalk');
 
 let packages = glob.sync(__dirname + '/packages/@react-{aria,spectrum,stately}/*/package.json');
 
+let errors = false;
+
+// soft assert won't fail the whole thing, allowing us to accumulate all errors at once
+// there's probably a nicer way to do this, but well... sometimes it's good enough. feel free to update me :)
+// maybe turn me into an actual eslint plugin so we get their format for styling
+function softAssert(val, message) {
+  try {
+    assert(val, message);
+  } catch {
+    console.error(chalk.red(message));
+    errors = true;
+  }
+}
+softAssert.deepEqual = function (val, val2, message) {
+  try {
+    assert.deepEqual(val, val2, message);
+  } catch {
+    console.error(chalk.red(message));
+    errors = true;
+  }
+}
+softAssert.equal = function (val, val2, message) {
+  try {
+    assert.equal(val, val2, message);
+  } catch {
+    console.error(chalk.red(message));
+    errors = true;
+  }
+}
+
 for (let pkg of packages) {
   let json = JSON.parse(fs.readFileSync(pkg));
-  assert(json.main, `${pkg} did not have "main"`);
-  assert(json.module, `${pkg} did not have "module"`);
-  assert(json.types, `${pkg} did not have "types"`);
-  assert(json.source, `${pkg} did not have "source"`);
-  assert.deepEqual(json.files, ['dist'], `${pkg} did not match "files"`);
-  assert.equal(json.sideEffects, false, `${pkg} is missing sideEffects: false`);
-  assert(!json.dependencies || !json.dependencies['@adobe/spectrum-css-temp'], `${pkg} has @adobe/spectrum-css-temp in dependencies instead of devDependencies`);
-  assert(json.dependencies && json.dependencies['@babel/runtime'], `${pkg} is missing a dependency on @babel/runtime`);
+  softAssert(json.main, `${pkg} did not have "main"`);
+  softAssert(json.module, `${pkg} did not have "module"`);
+  softAssert(json.types, `${pkg} did not have "types"`);
+  softAssert(json.source, `${pkg} did not have "source"`);
+  softAssert.deepEqual(json.files, ['dist'], `${pkg} did not match "files"`);
+  softAssert.equal(json.sideEffects, false, `${pkg} is missing sideEffects: false`);
+  softAssert(!json.dependencies || !json.dependencies['@adobe/spectrum-css-temp'], `${pkg} has @adobe/spectrum-css-temp in dependencies instead of devDependencies`);
+  softAssert(json.dependencies && json.dependencies['@babel/runtime'], `${pkg} is missing a dependency on @babel/runtime`);
 
   if (json.name.startsWith('@react-spectrum') && json.devDependencies && json.devDependencies['@adobe/spectrum-css-temp']) {
-    assert.deepEqual(json.targets, {
+    softAssert.deepEqual(json.targets, {
       main: {
         includeNodeModules: ['@adobe/spectrum-css-temp']
       },
@@ -26,5 +57,9 @@ for (let pkg of packages) {
     }, `${pkg} did not match "targets"`);
   }
 
-  assert(json.publishConfig && json.publishConfig.access === 'public', `${pkg} has missing or incorrect publishConfig`);
+  softAssert(json.publishConfig && json.publishConfig.access === 'public', `${pkg} has missing or incorrect publishConfig`);
+}
+
+if (errors) {
+  return process.exit(1);
 }
