@@ -1,14 +1,17 @@
-import classNames from 'classnames';
+import Asterisk from '@spectrum-icons/ui/Asterisk';
+import {classNames} from '@react-spectrum/utils';
+import {FieldLabelBase} from './types';
 import {filterDOMProps} from '@react-spectrum/utils';
-import {LabelProps} from '@react-types/label';
-import React, {forwardRef, ReactNode, RefObject} from 'react';
+import intlMessages from '../intl/*.json';
+import React, {forwardRef, RefObject} from 'react';
+import styles from '@adobe/spectrum-css-temp/components/fieldlabel/vars.css';
 import {useLabel} from '@react-aria/label';
+import {useMessageFormatter} from '@react-aria/i18n';
 
-interface SpectrumLabelBaseProps extends LabelProps {
-  labelClassName: string,
-  wrapperClassName: string,
-  componentName: string,
-  icon?: ReactNode 
+interface SpectrumLabelBaseProps extends FieldLabelBase {
+  labelClassName?: string,
+  wrapperClassName?: string,
+  componentName?: string
 }
 
 export const LabelBase = forwardRef((props: SpectrumLabelBaseProps, ref: RefObject<HTMLDivElement> & RefObject<HTMLLabelElement>) => {
@@ -16,9 +19,8 @@ export const LabelBase = forwardRef((props: SpectrumLabelBaseProps, ref: RefObje
   There are 3 cases:
   1. No children - only render the <label>, no wrapping div. `labelFor` required.
   Result: <label for="content">Label</label>
-  
   2. 1 child - render wrapping <div>. Automatically generate child `id` and label `for` attributes.
-  Result: 
+  Result:
     <label id="labelIdGenerated" for="contentIdGenerated">Label</label>
     <div>
       <input id="contentIdGenerated" aria-labelledby="labelIdGenerated" />
@@ -32,8 +34,8 @@ export const LabelBase = forwardRef((props: SpectrumLabelBaseProps, ref: RefObje
       <input id="input2" />
     </div>
   */
-  
-  let {    
+
+  let {
     label,
     children,
     className,
@@ -41,9 +43,19 @@ export const LabelBase = forwardRef((props: SpectrumLabelBaseProps, ref: RefObje
     wrapperClassName,
     labelFor,
     componentName,
-    icon,
+    necessityIndicator,
+    isRequired,
     ...otherProps
   } = props;
+
+  let formatMessage = useMessageFormatter(intlMessages);
+  let necessityLabel = isRequired ? formatMessage('(required)') : formatMessage('(optional)');
+  let icon = (
+    <Asterisk
+      className={classNames(styles, 'spectrum-FieldLabel-requiredIcon')}
+      size="S"
+      alt={formatMessage('(required)')} />
+    );
 
   let wrapper;
   let childArray = React.Children.toArray(children);
@@ -65,8 +77,8 @@ export const LabelBase = forwardRef((props: SpectrumLabelBaseProps, ref: RefObje
   if (!labelFor && childArray.length !== 1) {
     console.warn(`Missing labelFor attribute on ${componentName} with label "${label}"`);
   }
-  
   let fieldLabelClassName = classNames(
+    styles,
     labelClassName,
     childArray.length === 0 ? className : null
   );
@@ -74,20 +86,23 @@ export const LabelBase = forwardRef((props: SpectrumLabelBaseProps, ref: RefObje
   let fieldLabel = label ? (
     <label
       {...filterDOMProps(otherProps)}
-      className={fieldLabelClassName}
       {...labelAriaProps}
+      className={fieldLabelClassName}
       ref={ref}>
-      {label}
-      {icon && ' '}
-      {icon}
+      <span className={classNames(styles, 'spectrum-FieldLabel-label')}>
+        {label}
+      </span>
+      {necessityIndicator && ' '}
+      {necessityIndicator === 'label' && <span>{necessityLabel}</span>}
+      {necessityIndicator === 'icon' && isRequired && icon}
     </label>
   ) : (
     <div
-      className={fieldLabelClassName}
       {...filterDOMProps(otherProps)}
+      className={fieldLabelClassName}
       ref={ref} />
   );
-  
+
   if (childArray.length > 0) {
     if (wrapperClassName) {
       wrapper = (
