@@ -26,13 +26,49 @@ storiesOf('Form', module)
     )
   )
   .add(
-    'Label End',
+    'labelAlign: end',
     () => (
       <Form>
         <FormItem label="Company Title" labelAlign="end">
           <TextField placeholder="Company Title" />
         </FormItem>
         <FormItem label="Life Story" labelAlign="end">
+          <TextArea placeholder="Life Story" />
+        </FormItem>
+        <FormItem>
+          <Checkbox name="agreeToTerms"> 
+            <span>I agree to the <Link href="#" target="_self" style={{position: 'relative', zIndex: 1}}>terms and conditions</Link>.</span>
+          </Checkbox>
+        </FormItem>
+      </Form>
+    )
+  )
+  .add(
+    'isRequired: true/false, necessityLabel: label',
+    () => (
+      <Form>
+        <FormItem label="Company Title" isRequired necessityIndicator="label">
+          <TextField placeholder="Company Title" />
+        </FormItem>
+        <FormItem label="Life Story" necessityIndicator="label">
+          <TextArea placeholder="Life Story" />
+        </FormItem>
+        <FormItem>
+          <Checkbox name="agreeToTerms"> 
+            <span>I agree to the <Link href="#" target="_self" style={{position: 'relative', zIndex: 1}}>terms and conditions</Link>.</span>
+          </Checkbox>
+        </FormItem>
+      </Form>
+    )
+  )
+  .add(
+    'isRequired: true/false, necessityLabel: icon',
+    () => (
+      <Form>
+        <FormItem label="Company Title" isRequired necessityIndicator="icon">
+          <TextField placeholder="Company Title" />
+        </FormItem>
+        <FormItem label="Life Story" necessityIndicator="icon">
           <TextArea placeholder="Life Story" />
         </FormItem>
         <FormItem>
