@@ -1,4 +1,3 @@
-import _3DMaterials from '@spectrum-icons/workflow/3DMaterials';
 import {cleanup, render, within} from '@testing-library/react';
 import * as createId from '@react/react-spectrum/utils/createId';
 import {FormItem} from '../';
@@ -48,24 +47,22 @@ describe('FormItem', () => {
       ${'v2 FormItem multiple children'} | ${V2FormItem} | ${2}
     `('$Name should render a label if provided', ({Component, numChildren}) => {
       let tree = renderFormItem(Component, {label, labelFor}, numChildren);
-      let fieldLabel = tree.getByText(label);
+      let fieldLabel = tree.getByTestId(datatestid);
       expect(fieldLabel).toBeTruthy();
       expect(fieldLabel).toHaveAttribute('for', labelFor);
       expect(fieldLabel).toHaveAttribute('id', id);
+      expect(within(fieldLabel).getByText(label)).toBeTruthy();
     });
 
     it.each`
-      Name                               | Component     | numChildren
-      ${'v3 FormItem no children'}       | ${FormItem}   | ${0}
-      ${'v2 FormItem no children'}       | ${V2FormItem} | ${0}
-      ${'v3 FormItem one child'}         | ${FormItem}   | ${1}
-      ${'v2 FormItem one child'}         | ${V2FormItem} | ${1}
-      ${'v3 FormItem multiple children'} | ${FormItem}   | ${2}
-      ${'v2 FormItem multiple children'} | ${V2FormItem} | ${2}
-    `('$Name supports an icon with the label if provided', ({Component, numChildren}) => {
-      let tree = renderFormItem(Component, {label, labelFor, icon: <_3DMaterials />}, numChildren);
-      let fieldLabel = tree.getByText(label);
-      expect(within(fieldLabel).getByRole('img')).toBeTruthy();
+      Name                               | Component     | numChildren | props
+      ${'v3 FormItem no children'}       | ${FormItem}   | ${0}        | ${{isRequired: true, necessityIndicator: 'icon'}}
+      ${'v3 FormItem one child'}         | ${FormItem}   | ${1}        | ${{isRequired: true, necessityIndicator: 'icon'}}
+      ${'v3 FormItem multiple children'} | ${FormItem}   | ${2}        | ${{isRequired: true, necessityIndicator: 'icon'}}
+    `('$Name will show an asterix when required with icon indicator', ({Component, numChildren, props}) => {
+      let tree = renderFormItem(Component, {label, labelFor, ...props}, numChildren);
+      let fieldLabel = tree.getByTestId(datatestid);
+      expect(within(fieldLabel).getByRole('presentation')).toBeTruthy();
     });
 
     // Forwarding ref is v3 specific
@@ -77,7 +74,7 @@ describe('FormItem', () => {
     `('$Name should attach the user provided ref to the label', ({Component, numChildren}) => {
       let ref = React.createRef();
       let tree = renderFormItem(Component, {label, labelFor, ref}, numChildren);
-      let fieldLabel = tree.getByText(label);
+      let fieldLabel = tree.getByTestId(datatestid);
       expect(fieldLabel).toBe(ref.current);
     });
 
@@ -91,7 +88,7 @@ describe('FormItem', () => {
       ${'v2 FormItem multiple children'} | ${V2FormItem} | ${2}
     `('$Name should allow for additional dom props', ({Component, numChildren}) => {
       let tree = renderFormItem(Component, {label, labelFor, disabled: true}, numChildren);
-      let fieldLabel = tree.getByText(label);
+      let fieldLabel = tree.getByTestId(datatestid);
       expect(fieldLabel).toHaveAttribute('disabled', '');
     });
 
@@ -128,20 +125,6 @@ describe('FormItem', () => {
     it.each`
       Name                               | Component     | numChildren
       ${'v3 FormItem no children'}       | ${FormItem}   | ${0}
-      ${'v2 FormItem no children'}       | ${V2FormItem} | ${0}
-      ${'v3 FormItem one child'}         | ${FormItem}   | ${1}
-      ${'v2 FormItem one child'}         | ${V2FormItem} | ${1}
-      ${'v3 FormItem multiple children'} | ${FormItem}   | ${2}
-      ${'v2 FormItem multiple children'} | ${V2FormItem} | ${2}
-    `('$Name doesn\'t support an icon if label isn\'t provided', ({Component, numChildren}) => {
-      let tree = renderFormItem(Component, {labelFor, icon: <_3DMaterials />}, numChildren);
-      let fieldLabel = tree.getByTestId(datatestid);
-      expect(within(fieldLabel).queryByRole('img')).toBeFalsy();
-    });
-
-    it.each`
-      Name                               | Component     | numChildren
-      ${'v3 FormItem no children'}       | ${FormItem}   | ${0}
       ${'v3 FormItem one child'}         | ${FormItem}   | ${1}
       ${'v3 FormItem multiple children'} | ${FormItem}   | ${2}
     `('$Name should attach the user provided ref to the div', ({Component, numChildren}) => {
@@ -166,21 +149,6 @@ describe('FormItem', () => {
     });
   });
 
-  describe('with no children', () => {
-    it.each`
-      Name             | Component     | label
-      ${'v3 FormItem'} | ${FormItem}   | ${label}
-      ${'v2 FormItem'} | ${V2FormItem} | ${label}
-      ${'v3 FormItem'} | ${FormItem}   | ${null}
-      ${'v2 FormItem'} | ${V2FormItem} | ${null}
-    `('$Name should combine the provided class name and labelClassName', ({Component}) => {
-      let tree = renderFormItem(Component, {label, labelFor, className: 'testClass', labelClassName: 'labelClass'}, 0);
-      let fieldLabel = tree.getByText(label);
-      expect(fieldLabel).toHaveAttribute('class', expect.stringContaining('testClass'));
-      expect(fieldLabel).toHaveAttribute('class', expect.stringContaining('labelClass'));
-    });
-  });
-
   describe('with 1 child', () => {
     it.each`
       Name             | Component
@@ -193,10 +161,11 @@ describe('FormItem', () => {
       createIdMock.mockReturnValueOnce('second');
 
       let tree = renderFormItem(Component, {label}, 1);
-      let fieldLabel = tree.getByText(label);
+      let fieldLabel = tree.getByTestId(datatestid);
       expect(fieldLabel).toBeTruthy();
       expect(fieldLabel).toHaveAttribute('for', 'second');
       expect(fieldLabel).toHaveAttribute('id', 'first');
+
 
       let button = tree.getByTestId('testbutton');
       expect(button).toHaveAttribute('id', 'second');
@@ -212,17 +181,6 @@ describe('FormItem', () => {
       renderFormItem(Component, {label}, 1);
       expect(spyWarn).not.toHaveBeenCalled();
     });
-
-    it.each`
-      Name             | Component
-      ${'v3 FormItem'} | ${FormItem}
-      ${'v2 FormItem'} | ${V2FormItem}
-    `('$Name shouldn\'t combine the provided class name and labelClassName', ({Component}) => {
-      let tree = renderFormItem(Component, {label, className: 'testClass', labelClassName: 'labelClass'}, 1);
-      let fieldLabel = tree.getByText(label);
-      expect(fieldLabel).toHaveAttribute('class', expect.not.stringContaining('testClass'));
-      expect(fieldLabel).toHaveAttribute('class', expect.stringContaining('labelClass'));
-    });
   });
 
   describe('with more than 1 child', () => {
@@ -237,10 +195,11 @@ describe('FormItem', () => {
       createIdMock.mockReturnValueOnce('second');
 
       let tree = renderFormItem(Component, {label, labelFor}, 2);
-      let fieldLabel = tree.getByText(label);
+      let fieldLabel = tree.getByTestId(datatestid);
       expect(fieldLabel).toBeTruthy();
       expect(fieldLabel).toHaveAttribute('for', labelFor);
       expect(fieldLabel).toHaveAttribute('id', 'first');
+
 
       let firstButton = tree.getByTestId('testbutton');
       expect(firstButton).toBeTruthy();
@@ -250,17 +209,6 @@ describe('FormItem', () => {
       let secondButton = tree.getByTestId('testbutton2');
       expect(secondButton).not.toHaveAttribute('id');
       expect(secondButton).not.toHaveAttribute('aria-labelledby');
-    });
-
-    it.each`
-      Name             | Component
-      ${'v3 FormItem'} | ${FormItem}
-      ${'v2 FormItem'} | ${V2FormItem}
-    `('$Name shouldn\'t combine the provided class name and labelClassName', ({Component}) => {
-      let tree = renderFormItem(Component, {label, className: 'testClass', labelClassName: 'labelClass'}, 2);
-      let fieldLabel = tree.getByText(label);
-      expect(fieldLabel).toHaveAttribute('class', expect.not.stringContaining('testClass'));
-      expect(fieldLabel).toHaveAttribute('class', expect.stringContaining('labelClass'));
     });
   });
 });
