@@ -1,5 +1,5 @@
 import {action} from '@storybook/addon-actions';
-import Info from '@react/react-spectrum/Icon/Info';
+import Info from '@spectrum-icons/ui/InfoMedium';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {TextField} from '../';
@@ -27,13 +27,17 @@ storiesOf('TextField', module)
     () => render({isDisabled: true})
   )
   .add(
+    'isQuiet, isDisabled',
+    () => render({isDisabled: true, isQuiet: true})
+  )
+  .add(
     'validationState: invalid',
     () => render({validationState: 'invalid'})
   )
   .add(
     'validationState: invalid, validationTooltip',
     () => render({
-      validationState: 'invalid', 
+      validationState: 'invalid',
       validationTooltip: (
         <Tooltip>
           tooltip
@@ -64,6 +68,14 @@ storiesOf('TextField', module)
   .add(
     'icon: Info, isQuiet',
     () => render({icon: <Info />, isQuiet: true})
+  )
+  .add(
+    'icon: Info, isDisabled',
+    () => render({icon: <Info />, isDisabled: true})
+  )
+  .add(
+    'icon: Info, isQuiet, isDisabled',
+    () => render({icon: <Info />, isQuiet: true, isDisabled: true})
   )
   .add(
     'icon: Info, validationState: invalid, isQuiet',
