@@ -1,8 +1,12 @@
-import {TextFieldState} from '@react-types/textfield';
 import {useControlledState} from '@react-stately/utils';
 import {ValueBase} from '@react-types/shared';
 
-export function useTextFieldState(props: ValueBase<string | number | string[]>): TextFieldState {
+export interface SearchFieldState {
+  value: string,
+  setValue: (val: string, ...args: any) => void
+}
+
+export function useSearchFieldState(props: ValueBase<string | number | string[]>): SearchFieldState {
   let [value, setValue] = useControlledState(toString(props.value), toString(props.defaultValue) || '', props.onChange);
 
   return {
@@ -15,6 +19,6 @@ function toString(val) {
   if (val == null) {
     return;
   }
-    
+
   return val.toString();
 }
